VS-101 Start transform news images names
