--- client/src/types/news-item-type.ts
@@ -1,15 +1,0 @@
-export interface NewsItemType {
-    id: string;
-    datetime: string;
-    title_en: string;
-    title_ru: string;
-    subTitle_en: string;
-    subTitle_ru: string;
-    dir: string;
-    img_back: string;
-    img_backfull: string;
-    imagescount: number;
-    videoscount: number;
-    text_en: string;
-    text_ru: string;
-}
