+++ server/config.yaml
@@ -5,10 +5,9 @@ app:
 
   # Directory where images will be stored
 directories:
-  works_dir_save: "../client/content/works/"
+  works_dir_save: "d:/projects/vera-art/client/"
   works_db_dir_prefix: "/content/works/"
-  sale_dir_save: "../client/content/sales/"
+  sale_dir_save: "d:/projects/vera-art/client/"
   sale_db_dir_prefix: "/content/sale/"
-  news_dir_save: "../client/content/news/"
+  news_dir_save: "d:/projects/vera-art/client/"
   news_db_dir_prefix: "/content/news/"
-  news: "../client/src/assets/images/news/"
